feat(engine): on-screen mobile keyboard (#27)
* feat(engine): on-screen mobile keyboard

This PR adds a mobile-only on-screen virtual keyboard to the client. It
has minimal client edits, just some parts in `GameShell.ts` where we
were already mostly hooked in for mobile anyway.

It provides a far superior mobile typing experience - the existing code
would not even present a keyboard at all when in fullscreen!

It supports most of the symbols a regular keyboard could type,
specifically all of those that people would ordinarily use (I hope?)

It can be dragged around the screen, and shows/hides as appropriate when
an input field is touched.

* feat(engine): restrict clip

---------

Co-authored-by: locked-chest <179689881+sentientflare@users.noreply.github.com>